requirements.txt
pandas==2.2.3
openpyxl==3.1.5
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
rich==13.9.4
sqlglot==25.30.0
streamlit==1.41.1
pytest==8.3.4
openai>=1.59.0
pydantic==2.10.3
python-dateutil==2.9.0.post0
cryptography>=41.0.0
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.27.2
pyyaml==6.0.2